requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cache==1.0.3
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.0.6
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.10
Markdown==3.7
numpy==2.2.4
packaging==24.1
pandas==2.2.3
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-c==3.2.9
psycopg2-binary==2.9.10
pyactiveresource==2.2.2
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
ShopifyAPI==12.7.0
six==1.17.0
sqlparse==0.5.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wheel==0.45.1
whitenoise==6.7.0
psycopg[c]
redis==5.0.1
django-redis==5.4.0